Add a debug function for detecting API status
This commit can be ignored, I just wasn't really sure how to test the Boss Manager in Unity, so I added a method to let the Boss Manager let KMBossModule know when it's loaded.
Unfortunately, most of the work I did here was done in KMBossModule.cs, which isn't included in this project.
It had four settings: Unavailable (Boss Manager not installed), Loading (Either in Refresh corotuine or the _functions dictionary is null), Unsupported (earlier version of Boss Manager that does not have the key), and Loaded.
I can provide the new KMBossModule I used for this if you think it could be helpful.